refactor: renombrar hooks de gestión de tipos de habitación para mayor claridad
